requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==40.0.0
Django==4.1.4
django-ratelimit==4.0.0
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
mysqlclient==2.1.1
Pillow==9.4.0
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7
redis==4.5.1
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
